Move shutdown out, test launch and kill.
